+++ ClientSocket.h
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 
 #ifdef _MSC_VER
-typedef size_t ssize_t;
+typedef int ssize_t;
 #else
 #include <unistd.h>
 #endif
